% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleGen.R
\name{filter_rules}
\alias{filter_rules}
\title{Filter out duplicate rules}
\usage{
filter_rules(rules, x_df, filtering_tol = 0, priority = NULL, verbose = F)
}
\arguments{
\item{rules}{A character vector containing the rules.}

\item{x_df}{a dataframe containing the observations to use when computing the rules.
The dataframe should not contain the outcome \eqn{y}.}

\item{filtering_tol}{By how many points can rules differ in terms of support and
still be considered duplicates. By default, \code{filtering_tol=0} implies that
only rules that coincide on each point of the dataset.}

\item{priority}{numerical vector of priorities: one value per rule. A higher value
means a more desirable rule. If left to NULL, the priority is given based on the order
of appearance in the vector.}

\item{verbose}{Logical value denoting whether a progress bar should be printed to show
the progress of the filtering.}
}
\value{
A character vector containing all the rules that remained after filtering.
}
\description{
This function filters out rules that are duplicates in the sense that their support
coincides on a given dataset. Rules can be given a priority when choosing which one to delete.
}
