% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{gendata.friedman1}
\alias{gendata.friedman1}
\title{Simulate data using the Friedman generating function}
\usage{
gendata.friedman1(n, p = 100, rho = NULL, sd = 1, binary = F)
}
\arguments{
\item{n}{number of observations that should be generated.}

\item{p}{number of predictors that should be generated. Should be at least 5. For any
number higher than \eqn{p=5}, the remaining features are all noise features.}

\item{rho}{correlation induced across the features.}

\item{sd}{In the case of continuous outcome, the standard deviation of the normally-distributed
irreducible noise. For binary outcome, this parameter is ignored.}

\item{binary}{A logical parameter determining whether the outcome should binarized (\code{TRUE})
or should be left continuous (\code{FALSE}).}
}
\value{
A dataframe containing the synthetic dataset.
}
\description{
This function creates synthetic data with the Friedman data-generating scheme:
\deqn{x_1,\ldots,x_{5} \sim \mathcal{U}(0,1),}
\deqn{\epsilon \sim \mathcal{N}(0,\sigma^2),}
\deqn{y=10\sin(\pi x_1x_2)+20(x_3-0.5)^2+10x_4+5x_5+\epsilon.}
This function may also binarize the outcome in the following manner:
\deqn{\eta=10\sin(\pi x_1x_2)+20(x_3-0.5)^2+10x_4+5x_5,}
\deqn{p=\frac{exp(\eta-\overline{\eta})}{exp(\eta-\overline{\eta})+1},}
\deqn{y \sim \mathcal{B}\textbf{e}(p),}
with \eqn{\overline{\eta} \approx 14.4} computed as the approximate mean of
\eqn{\eta} under this data-generation scheme, in the case of independent features.
Features may be generated to be correlated, using a gaussian copula.
}
