% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{hs}
\alias{hs}
\title{Fit a horseshoe model}
\usage{
hs(
  y,
  X,
  family = c("gaussian", "binomial"),
  p.lin = ncol(X),
  method.tau.rules = c("halfCauchy", "truncatedCauchy", "fixed"),
  tau.rules = 1,
  method.tau.lin = c("halfCauchy", "truncatedCauchy", "fixed"),
  tau.lin = 1,
  method.tau = c("halfCauchy", "truncatedCauchy", "fixed"),
  tau = 1,
  method.sigma = c("Jeffreys", "fixed"),
  Sigma2 = 1,
  burn.in = 1000,
  nmc = 5000,
  thin = 1,
  prior_lambda = NULL,
  Beta_init = rep(0, ncol(X)),
  Lambda_init = rep(1, ncol(X)),
  ping = 1000,
  verbose = F
)
}
\arguments{
\item{y}{a numeric vector of outcomes used as training data to fit the model. For binary outcome, this should
be converted into zeroes and ones.}

\item{X}{a numeric \eqn{n \times p} matrix whose columns are predictors and whose rows are observations.
This is used as training data to fit the model.}

\item{family}{A character string denoting the type of (Generalized) linear model
that is being fitted}

\item{p.lin}{number of columns of X which are linear terms and should thus get their
own global shrinkage separately from the last remaining columns of X. By default, all
columns are linear terms and therefore all columns share the same shrinkage.}

\item{method.tau.rules}{character string denoting the prior distribution of the
global shrinkage assigned to rules: the choice is between a half-Cauchy prior
(\code{"halfCauchy"}),a truncated Cauchy prior (\code{"truncatedCauchy"}) or
a fixed value (\code{"fixed"})}

\item{tau.rules}{a numeric entry denoting which value the global shrinkage for rules should take,
in the case where \code{method.tau.rules} is set to \code{"fixed"}. If other values of \code{method.tau.lin}
are chosen, this parameter is used as initial value to start the Gibbs sampling from.}

\item{method.tau.lin}{character string denoting the prior distribution of the
global shrinkage assigned to linear terms: the choice is between a half-Cauchy prior
(\code{"halfCauchy"}),a truncated Cauchy prior (\code{"truncatedCauchy"}) or
a fixed value (\code{"fixed"})}

\item{tau.lin}{a numeric entry denoting which value the global shrinkage for linear terms should take,
in the case where \code{method.tau.lin} is set to \code{"fixed"}. If other values of \code{method.tau.lin}
are chosen, this parameter is used as initial value to start the Gibbs sampling from.}

\item{method.tau}{character string denoting the prior distribution of the
overall global shrinkage assigned to all terms: the choice is between a half-Cauchy prior
(\code{"halfCauchy"}),a truncated Cauchy prior (\code{"truncatedCauchy"}) or
a fixed value (\code{"fixed"})}

\item{tau}{a numeric entry denoting which value the global shrinkage for linear terms should take,
in the case where \code{method.tau} is set to \code{"fixed"}. If other values of \code{method.tau}
are chosen, this parameter is used as initial value to start the Gibbs sampling from.}

\item{method.sigma}{character string denoting which prior distribution the error variance \eqn{\sigma^2}
should have: the choice is between Jeffre's prior (\code{"Jeffreys"}) or a fixed
value (\code{"fixed"}).}

\item{Sigma2}{a numeric entry denoting which value the error variance \eqn{\sigma^2} should take,
in the case where \code{method.sigma} is set to \code{"fixed"}. If other values of \code{method.sigma}
are chosen, this parameter is ignored.}

\item{burn.in}{Number of burn-in MCMC iterations that are discarded before
the draws start to be recorded.}

\item{nmc}{Number of MCMC iterations after the initial burn-in period}

\item{thin}{Thinning parameter: if saving all MCMC draws is computationally
infeasible, thinning makes the model only record every \code{thin}-th iteration.}

\item{prior_lambda}{a vector of \eqn{p} positive numeric entries representing local shrinkage relaxation
(\code{prior_lambda > 1}) or contraction (\code{prior_lambda < 1}). The \eqn{j}-th entry of this vector
denotes by how much the local shrinkage should be relaxed for the \eqn{j}-th predictor.}

\item{Beta_init}{a numeric vector of lenth \eqn{p} representing the initial coefficients to start
the Gibbs sampler with.}

\item{Lambda_init}{a numeric vector of lenth \eqn{p} representing the initial values of
local shrinkage to start the Gibbs sampler with.}

\item{ping}{If the model is selected to be verbose, this denotes every how many
MCMC draws an update message should be printed}

\item{verbose}{Logical value determining whether updates should be given on the progress}
}
\value{
BetaSamples A matrix where every column is a MCMC sample of the coefficients.

BetaHat A vector containing the posterior mean of the coefficients.

TauLinSamples A vector where every entry is a MCMC sample of the global shrinkage of the linear terms.

TauLinHat The posterior mean of the global shrinkage of the linear terms.

TauRulesSamples A vector where every entry is a MCMC sample of the global shrinkage of the rules.

TauRulesHat The posterior mean of the global shrinkage of the rules.

LambdaSamples A matrix where every column is a MCMC sample of the local shrinkages, one entry per term.

LambdaHat A vector containing the posterior mean of the local shrinkages.
}
\description{
This function fits a plain horseshoe Bayesian regression for continuous and binary data with
a MCMC algorithm based on Gibbs sampling. This function is also compatible with a grouped
local shrinkage for linear terms and rule terms separately.
}
