% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleGen.R
\name{PRF}
\alias{PRF}
\title{Generate rules with a Parametric Random Forest}
\usage{
PRF(x_df, y, ntree = 500, maxdepth = 3, rounding.digits = 3, disaggregate = T)
}
\arguments{
\item{x_df}{a dataframe containing the observations to use when computing the rules.
The dataframe should not contain the outcome \eqn{y}.}

\item{y}{a numerical vector representing the outcome \eqn{y}.}

\item{ntree}{number of trees that should be used when generating the rules}

\item{maxdepth}{maximum depth of the trees used to generate the rules. In particular,
this induces a maximum on the order of interactions between predictors.}

\item{rounding.digits}{Number of decimals that the splitting points
are rounded to in the rule generation. This rounding simplifies readability
of rules and prevents rules from being very similar but not identical.}

\item{disaggregate}{a logical value determining whether rules should be extracted
by also considering all possible subrules (\code{TRUE}) or if rules should only
be extracted in a top-down manner as in RuleFit and HorseRule (\code{FALSE}).}
}
\value{
A character vector containing all the rules that were generated. The rules
are not filtered for duplicates. To filter rules, check the function \code{filter_rules}.
}
\description{
This function fits a Parametric Random Forest to generate dychotomous rules that
may be used to fit a RuleSHAP model. This is designed for a continuous outcome.
}
