% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleGen.R
\name{RF2unified}
\alias{RF2unified}
\title{Convert Random Forest model to data frame format}
\usage{
RF2unified(rf, data, rounding.digits = 3)
}
\arguments{
\item{rf}{A Random Forest model as fitted from the \code{randomForest} package.}

\item{data}{a data frame containing the observations to use when computing the rules.
The data frame should not contain the outcome \eqn{y}.}

\item{rounding.digits}{Number of decimals that the splitting points
are rounded to in the rule generation. This rounding simplifies readability
of rules and prevents rules from being very similar but not identical.}
}
\value{
A data frame expressing the tree ensemble as a sequence of decision rules.
}
\description{
This function converts a Random Forest fit into a data frame format compatible with rule manipulation
}
