% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleGen.R
\name{RuleMats}
\alias{RuleMats}
\title{Compute matrices that identify the rules}
\usage{
RuleMats(rules, x_df)
}
\arguments{
\item{rules}{a character vector containing the rules to compute the matrices for}

\item{x_df}{a dataframe containing the observations to compute the matrices for}
}
\value{
RulePredMat A matrix with as many rows as there are rules and as many
columns as there are predictors. The \eqn{j}-th column has ones on the entries
corresponding to rules that involve the \eqn{j}-th predictor, while all
remaining entries are zeroes. Can be computed with the \link{RuleMats} function.

Rs a list of as many matrices as there are rules. The \eqn{j}-th element
is a matrix corresponding to the \eqn{j}-th rule. Each of its columns corresponds
to the 0-1 encoding of a subrule of the \eqn{j}-th rule, as observed in the data
provided with input parameter \code{x}.
}
\description{
This function computes the matrices that are needed to compute marginal
Shapley values of a RuleSHAP model. These matrices check which points satisfy
each of the subrules of each rule. They also check which predictors are involved
in the definition of each rule.
}
