% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleGen.R
\name{unified2rules}
\alias{unified2rules}
\title{Convert a unified dataframe to a vector of rules}
\usage{
unified2rules(unified, maxdepth = 3, disaggregate = F)
}
\arguments{
\item{unified}{A dataframe containing all essential information about the tree ensemble.
Can be produced with the \code{*.unify} functions from the \code{treeshap} package,
or from the \code{RF2unified} and \code{XGB2unified} functions in this package.}

\item{maxdepth}{maximum depth of the trees used to generate the rules. In particular,
this induces a maximum on the order of interactions between predictors.}

\item{disaggregate}{a logical value determining whether rules should be extracted
by also considering all possible subrules (\code{TRUE}) or if rules should only
be extracted in a top-down manner as in RuleFit and HorseRule (\code{FALSE}).}
}
\value{
A character vector containing all the rules extracted from the tree ensemble.
}
\description{
This function takes a tree ensemble as input and extracts all the rules that define it.
}
